Modularize the built-in user manual module to a separate crate.
