+++ sic_user_manual/src/lib.rs
